Layout_TestCase/http/tests/xmlhttprequest/access-control-preflight-credential-sync-expected.txt

ALERT: http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/cookies/resources/cookie-utility.php?queryfunction=setFooCookie
ALERT: XHR response - Set the foo cookie
Test case for bug 37781: [XHR] Cross-Origin synchronous request with credential raises NETWORK_ERR

PASSED
